Added Ability To Have Delegate Perform Interactions
Refactored Pages Request Perform interaction code to allow a different world object to be used in its place, allowing a page to for example set a key as its delegate, then have the key interact with a door unlocking it.
Also added several custom cursors, and code in pagecontroller to set the cursor.
Refactored interactions is several classes to remove redundancies